+++ Example/example_main.tf
@@ -1,70 +1,57 @@
 #
-# Copyright 2023 Emmanuel Bergerat
+# Copyright 2024 Emmanuel Bergerat
 #
 
 #--------------------------------------------------------------
-#   Test #{MODULEDISPLAYNAME}# module Main
+#   Test Key vault module Main
 #--------------------------------------------------------------
-# Test naming cascade
-module "rg_cascade" {
+#   / Resource Group
+module "rg" {
+  # Terraform Cloud use
+  source  = "app.terraform.io/embergertf/resourcegroup/azurerm"
+  version = "~> 2.1"
+
+  # Name override
+  # name_override = var.name_override
+
+  # Naming convention
+  naming_values = var.naming_values
+  # region_code     = var.region_code
+  # subsc_code      = var.subsc_code
+  # env             = var.env
+  # base_name       = var.base_name
+  # additional_name = var.additional_name
+  # iterator        = var.iterator
+  # owner           = var.owner
+
+  # # Random
+  # add_random = var.add_random
+  # rnd_length = var.rnd_length
+
+  # additional_tags = null
+}
+
+module "publicip" {
+  source  = "app.terraform.io/embergertf/publicip/http"
+  version = "~> 1.0"
+}
+
+
+module "kv_module_localtest" {
   # Local use
-  source = "../../terraform-azurerm-base"
+  source = "../../terraform-azurerm-keyvault"
 
-  region_code     = "usnc"
-  subsc_code      = "azint"
-  env             = "dev"
-  base_name       = "basemodule"
-  additional_name = ""
-  iterator        = "01"
-  owner           = "john.doe@internet.com"
-  additional_tags = {
-    app_id  = "1"
-    test_by = "emberger"
-  }
+  # Naming convention
+  naming_values = module.rg.naming_values
+  add_random    = true
 
-  # Resource naming inputs
-  resource_type_code = "rg"
-  max_length         = 64
-  no_dashes          = false
-  add_random         = true
-  rnd_length         = 2
-}
+  # Key vault settings
+  resource_group_name = module.rg.resource_group_name
+  tenant_id           = data.azurerm_client_config.current.tenant_id
 
-# Test the base module outputs by creating a Resource Group
-resource "azurerm_resource_group" "this" {
-  name     = module.rg_cascade.name
-  location = module.rg_cascade.location
 
-  tags = merge(module.rg_cascade.tags, {
-    random_suffix = module.rg_cascade.random_suffix
-  })
-}
+  public_internet_ips_to_allow = [module.publicip.public_ip]
+  deploy_validation_secret     = var.deploy_validation_secret
 
-# Test name override
-module "rg_override" {
-  # Local use
-  source = "../../terraform-azurerm-base"
-
-  name_override = "biglittleoverride"
-  additional_tags = {
-    app_id  = "1"
-    test_by = "emberger"
-  }
-
-  # Resource naming inputs
-  resource_type_code = "rg"
-  max_length         = 64
-  no_dashes          = false
-  add_random         = true
-  rnd_length         = 2
-}
-
-# Test the base module outputs by creating a Resource Group
-resource "azurerm_resource_group" "or" {
-  name     = module.rg_override.name
-  location = module.rg_override.location
-
-  tags = merge(module.rg_override.tags, {
-    random_suffix = module.rg_cascade.random_suffix
-  })
+  additional_tags = var.kv_additional_tags
 }