+++ CHANGELOG.md
@@ -13,8 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed -->
 
-## [1.0.0] - #{CREATIONDATE}# - #{MODULEDISPLAYNAME}# module 1st Release
+## [1.0.0] - 2024-11-11 - Key vault module Initial Release
 
 ### Added
 
-- Added #{MODULEDISPLAYNAME}# module
+- Added Key vault module